requirements.txt
asgiref==3.5.2
decouple==0.0.7
Django==4.1.3
django-filter==22.1
django-versatileimagefield==2.2
djangorestframework==3.14.0
Pillow==9.4.0
python-decouple==3.6
python-magic==0.4.27
pytz==2022.6
sqlparse==0.4.3
